monitor: Fix integer overruns with EIR parsing
For tracking how much has been parsed of the EIR we must use a greater
than uint8 integer to avoid having a large (e.g. 255) field length cause
our counter to wrap around and start from zero.

+++ monitor/packet.c
@@ -2231,7 +2231,7 @@ static const struct {
 
 static void print_eir(const uint8_t *eir, uint8_t eir_len, bool le)
 {
-	uint8_t len = 0;
+	uint16_t len = 0;
 
 	if (eir_len == 0)
 		return;
